Fix contact form placement/sizing on production
This one is weird. On my local, it looks fine. On prod, additional
styles are added to the form that I can't override in the DOM without
adding !important, making the table 27% wide and floated right. Regardless,
there's no need for concern because this is unique to the contact page's contact form

Fixes DD 71427-79