Prevent 'Sign In/Create Account' text from changing and selecting all products
This is a WEIRD one... Somehow the target is just [role="button"], which exists on other
parts of the page, so if you clicked 'select all' in the related products sidebar, it would change
the Sign In/Create Account text to Select All / Deselect All. Same goes for the country switcher.

And if you clicked the Sign In/Create Account or country switcher links, it would select all
of the products in the related products sidebar. Weird.

Fixes DD 71427 425