Adjust JS placement
This probably won't do anything, but I'm hoping it does. On my local,
these fire 100% of the time and the brands page loads with the styled nav.
On prod, these hardly ever run and so the brands nav is only styled ~10% of
the time. I have yet to see it fail on my local.

Fixes DD 71427 346